refactor(governance): extract GovernanceDeliveryService from admin (admin-service-split-delivery SDD)
Split the delivery-orchestration domain out of the 1012-line
GovernanceAdminService into a new GovernanceDeliveryService
(the architecture 1000-line cap test had gone red on admin_service).

Migrated (verbatim, behavior-preserving) — deliver_pending /
deliver_by_worker / _run_delivery / create_and_send_reminder (~436 lines):
- delivery_service.py: new GovernanceDeliveryService with the 4 methods;
  __init__ injects notify_repo/audit_repo/task_repo/config/notify_sender/
  render_svc/lifecycle_svc (reminder chain advance via lifecycle.refresh_snapshot).
- service_protocols.py + api re-export: GovernanceDeliveryServiceProtocol
  (runtime_checkable); drop the migrated methods from
  GovernanceAdminServiceProtocol.
- admin_router.py: tickets:deliver / tickets:remind / scan-and-deliver
  (deliver phase) inject _DeliverySvc and call it directly — no proxy shell;
  scan-and-deliver cron-tick phase still uses scan_svc.
- admin_service.py: 1012→583 lines, drops notify_sender injection (was a dead
  param after the move); render_svc kept (force_renew_with_record renders
  first_send body). docstring updated to reflect the split.
- economy_governance_module.py: _delivery_service provider + protocol binding;
  admin provider drops notify_sender.
- tests: new test_delivery_service.py contract tests; TestDeliverByWorker +
  coverage_supplement deliver-pending cases repointed to GovernanceDeliveryService.

Verification: 857 governance+endpoint+arch tests green, arch cap now green
(admin 583<1000); variant-line coverage delivery ~89% / admin ~97%
(uncovered lines are pre-existing error/degradation/audit-fail branches,
not newly introduced). E2E DSL (live backend): 35 pass / 1 fail — the lone
fail (TC-C05) asserts admin_close sets no cooldown, a stale expectation from
the separate cooldown-drift line (admin_close semantics moved to set
cooldown_days); unrelated to this refactor (admin_close lives in untouched
workflow_service).
