+++ src/backend/src/agentclaw/community/api/governance_service.py
@@ -20,6 +20,7 @@ from typing import TYPE_CHECKING, Any, Protocol, runtime_checkable
 from agentclaw.community.core.economy.governance.services.service_protocols import (
     GovernanceAdminServiceProtocol,
     GovernanceAuditReadServiceProtocol,
+    GovernanceDeliveryServiceProtocol,
     GovernanceLifecycleServiceProtocol,
     GovernanceWhitelistServiceProtocol,
     GovernanceWorkflowServiceProtocol,
