+++ src/backend/tests/community/core/economy/governance/test_admin_service.py
@@ -35,6 +35,9 @@ from agentclaw.community.core.economy.governance.repositories.whitelist_repo imp
 from agentclaw.community.core.economy.governance.services.admin_service import (
     GovernanceAdminService,
 )
+from agentclaw.community.core.economy.governance.services.delivery_service import (
+    GovernanceDeliveryService,
+)
 from agentclaw.community.core.economy.governance.services.notify_render_service import (
     NotifyRenderService,
 )
@@ -147,11 +150,39 @@ def _build_svc(engine):
         audit_repo=audit_repo,
         task_repo=task_repo,
         config=FakeGovernanceConfig(),
-        notify_sender=FakeNotifySender(),
         lifecycle_svc=lifecycle_svc,
         render_svc=NotifyRenderService(),
     )
     return svc, db, cache
+
+
+def _build_delivery_svc(engine):
+    """Build delivery service with in-memory DB + fake notify sender.
+
+    投递编排域从 admin_service 抽出(admin-split SDD)后的专用构造器,复用与
+    ``_build_svc`` 相同的 repo/lifecycle/render 注入,仅返回
+    GovernanceDeliveryService(admin_service 不再持有投递方法)。
+    """
+    Session = sessionmaker(bind=engine, expire_on_commit=False)
+    db = FakeDB(lambda: Session(bind=engine))
+    notify_repo = NotifyLogRepository(db=db)
+    task_repo = TaskRecordRepository(db=db)
+    audit_repo = GovernanceAuditRepository(db=db)
+    lifecycle_svc = GovernanceLifecycleService(
+        task_repo=task_repo,
+        notify_repo=notify_repo,
+        audit_repo=audit_repo,
+    )
+    svc = GovernanceDeliveryService(
+        notify_repo=notify_repo,
+        audit_repo=audit_repo,
+        task_repo=task_repo,
+        config=FakeGovernanceConfig(),
+        notify_sender=FakeNotifySender(),
+        render_svc=NotifyRenderService(),
+        lifecycle_svc=lifecycle_svc,
+    )
+    return svc, db, None
 
 
 def _build_workflow_svc(engine):
@@ -834,7 +865,7 @@ class TestDeliverByWorker:
 
     def test_dry_run_builds_without_sending(self, session, engine):
         """dry_run=true:只构建投递清单,不发钉钉、不更新 notify_status、不写审计。"""
-        svc, _, _ = _build_svc(engine)
+        svc, _, _ = _build_delivery_svc(engine)
         _make_notification(
             session, notification_id="n-a1", worker_id="ownerA:botA",
             bot_id="botA", owner_id="ownerA", notify_status="pending",
@@ -861,7 +892,7 @@ class TestDeliverByWorker:
 
     def test_live_send_updates_status_and_audits(self, session, engine):
         """非 dry_run:发送成功 → notify_status=sent + 审计 NOTIFICATION_SENT。"""
-        svc, db, _ = _build_svc(engine)
+        svc, db, _ = _build_delivery_svc(engine)
         _make_notification(
             session, notification_id="n-live", worker_id="ownerA:botA",
             bot_id="botA", owner_id="ownerA", notify_status="pending",
@@ -888,7 +919,7 @@ class TestDeliverByWorker:
 
     def test_worker_id_filter_excludes_other_workers(self, session, engine):
         """只投递指定 worker 的 pending,其它 worker 的不动。"""
-        svc, _, _ = _build_svc(engine)
+        svc, _, _ = _build_delivery_svc(engine)
         _make_notification(
             session, notification_id="n-a", worker_id="ownerA:botA",
             bot_id="botA", owner_id="ownerA", notify_status="pending",
@@ -912,7 +943,7 @@ class TestDeliverByWorker:
 
     def test_empty_worker_returns_zero(self, session, engine):
         """该 worker 无 pending → total=0,sent_count=0,results 空。"""
-        svc, _, _ = _build_svc(engine)
+        svc, _, _ = _build_delivery_svc(engine)
         _make_notification(
             session, notification_id="n-x", worker_id="ownerX:botX",
             bot_id="botX", owner_id="ownerX", notify_status="pending",
@@ -928,7 +959,7 @@ class TestDeliverByWorker:
 
     def test_no_override_recipient_falls_back_to_owner(self, session, engine):
         """不传 override_recipient → 逐条按通知 owner 兜底发送(FakeNotifySender 记录)。"""
-        svc, _, _ = _build_svc(engine)
+        svc, _, _ = _build_delivery_svc(engine)
         _make_notification(
             session, notification_id="n-fb", worker_id="ownerA:botA",
             bot_id="botA", owner_id="ownerA", notify_status="pending",
@@ -945,7 +976,7 @@ class TestDeliverByWorker:
 
     def test_updates_ticket_delivery_status_on_success(self, session, engine):
         """发送成功后应更新 task_record 的 delivery_status 为 {notify_type}:sent。"""
-        svc, db, _ = _build_svc(engine)
+        svc, db, _ = _build_delivery_svc(engine)
         # 创建工单和 first_send 通知
         ticket_id = "t-deliver-status"
         notification_id = "n-deliver-status"
@@ -1016,16 +1047,14 @@ class TestDeliverByWorker:
             config=FakeGovernanceConfig(),
             lifecycle_svc=lifecycle_svc,
         )
-        svc = GovernanceAdminService(
-            cache=cache,
-            whitelist_service=whitelist_service,
+        svc = GovernanceDeliveryService(
             notify_repo=notify_repo,
             audit_repo=audit_repo,
             task_repo=task_repo,
             config=FakeGovernanceConfig(),
             notify_sender=FailingNotifySender(),  # 注入失败 sender
-            lifecycle_svc=lifecycle_svc,
             render_svc=NotifyRenderService(),
+            lifecycle_svc=lifecycle_svc,
         )
 
         # 创建工单和 first_send 通知
